Refactor math types:
- TPointX as float/double templates
- All "lightweight" math types moved to its own .cpp
