+++ libs/math/include/mrpt/math/TPolygon3D.h
@@ -9,6 +9,7 @@
 #pragma once
 
 #include <mrpt/math/TPoseOrPoint.h>
+#include <vector>
 
 namespace mrpt::math
 {
